requirements.txt
Django==4.0.6
django-jazzmin==2.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
Pillow
django-cors-headers
# para autenticar con google
django-allauth
# para subir imagenes en base 64
drf-extra-fields
